Set maximum python version to `3.9`.
A dependency `frozendict` (itself a dependency of `cairo-lang`) uses
deprecated syntax that was removed in python 3.10, and now raises an
`AttributeError`. In particular, it tries to import `Mapping` from
`collections`. Thus, this commit explicitly drops support for 3.10.

This commit also fixes the CI to reflect this change. Note that although
the CI workflow gives 3.7 as the version prior to this revision, the
workflow output suggests it may have been running under 3.10.

* Add `poetry env use` commands to `ci.yml` to fix python versioning mismatch.